added stuff to use BinanceAnalyze.js
