+++ Lib/toTheMinute.js
@@ -1,0 +1,4 @@
+module.exports = function(date) {
+    let coeff = 1000 * 60;
+    return new Date(Math.round(date.getTime() / coeff) * coeff)
+}
