+++ gdax-pulse.js
@@ -1,62 +1,30 @@
+const AnalyzeGdax = require('./AnalyzeFunctions/GdaxAnalyze')
+const toTheMinute = require('./Lib/toTheMinute')
+const AnalyzeBinance = require('./AnalyzeFunctions/BinanceAnalyze')
 class Pulse {
-    constructor(delay) {
+    constructor(delay,exchange) {
+        this.exchange = exchange === undefined ? "gdax" : exchange.toLowerCase();
         let now = new Date();
         this.lastMinute = now.getMinutes();
         this.lastHour = now.getHours();
         this.lastDay = now.getDate();
         this.lasUtcDay = now.getUTCDate();
         this.delay = delay !== undefined ? delay : 0;
+        if (this.exchange == "gdax"){
+            this.analyze = AnalyzeGdax.bind(this);
+        }
+        else if(this.exchange == "binance"){
+            this.analyze = AnalyzeBinance.bind(this);
+        }
+        else
+            throw new error("incorrect exchange, select gdax or binance")
         this.currentData = {
             time: toTheMinute(now),
             price: 0
         }
     }
 
-    analyze(message) {
-        //update Time
-        if (message.type === 'match') {
-            let priceOfMatch = parseFloat(message.price);
-            if (priceOfMatch !== this.currentData.price) {
-                this.currentData.price = priceOfMatch;
-                this.onNewPrice(this.currentData.price, this.currentData.time);
-            }
-        }
-
-        if (message.time) {
-            let current = new Date(message.time);
-            let now = new Date(current.getTime() - 1000 * this.delay)
-            if (this.lastMinute !== now.getMinutes()) {
-                this.lastMinute = now.getMinutes();
-                //update time
-                this.currentData.time = toTheMinute(now);
-                this.onMin1(this.currentData.price, this.currentData.time);
-                if (this.lastMinute % 5 === 0) {
-                    this.onMin5(this.currentData.price, this.currentData.time)
-                }
-                if (this.lastMinute % 15 === 0) {
-                    this.onMin15(this.currentData.price, this.currentData.time)
-                }
-
-                if (this.lastHour !== now.getHours()) {
-                    this.lastHour = now.getHours();
-                    this.onHour1(this.currentData.price, this.currentData.time);
-                    if (this.lastHour % 6 === 0) {
-                        this.onHour6(this.currentData.price, this.currentData.time)
-                    }
-
-                    if (this.lastDay !== now.getDate()) {
-                        this.lastDay = now.getDate();
-                        this.onDay(this.currentData.price, this.currentData.time)
-                    }
-
-                    if (this.lasUtcDay !== now.getUTCDate()) {
-                        this.lasUtcDay = now.getUTCDate();
-                        this.onUtcDay(this.currentData.price, this.currentData.time);
-                    }
-                }
-            }
-        }
-    }
+    analyze(message){}
 
     on(type, func) {
         if (type === 'm1') {
@@ -89,8 +57,3 @@ class Pulse {
 }
 
 module.exports = Pulse;
-
-function toTheMinute(date) {
-    let coeff = 1000 * 60;
-    return new Date(Math.round(date.getTime() / coeff) * coeff)
-}
