+++ package.json
@@ -26,6 +26,7 @@
     "mocha": "^5.2.0"
   },
   "dependencies": {
-    "gdax-sim": "^1.0.1"
+    "gdax-sim": "^1.0.1",
+    "node-binance-api": "^0.9.2"
   }
 }
